"working" node streams

+++ Assets/Scripts/Node.cs
@@ -22,7 +22,7 @@ public class Node : MonoBehaviour
 
     public List<Node> connectedNodes;
     public NodeType type;
-    public int influence;
+    public float influence;
     public float connectionRadius;
     public Node influenceSource;
 
@@ -115,21 +115,21 @@ public class Node : MonoBehaviour
 
     public void CheckPower()
     {
-        if (influence >= 10 && type != NodeType.Misinformed)
+        if (influence >= 10f && type != NodeType.Misinformed)
         {
             // This node is now influenced by bad information. RIP
             ChangeNodeType(NodeType.Misinformed);
-            influence = 10;
+            influence = 10f;
         }
 
-        if (influence <= -10 && type != NodeType.Reliable)
+        if (influence <= -10f && type != NodeType.Reliable)
         {
             // We have become a spreader of truths.
             ChangeNodeType(NodeType.Reliable);
-            influence = -10;
+            influence = -10f;
         }
 
-        if (type is NodeType.Misinformed or NodeType.Reliable && influence is < 5 and > -5)
+        if (type is NodeType.Misinformed or NodeType.Reliable && influence is < 5f and > -5f)
         {
             // Slowly changing our ways
             ChangeNodeType(NodeType.Neutral);
